Refactor ResumeReviewTable to its own file
